pyproject.toml
[tool.poetry]
name = "test_demo"
version = "0.1.0"
description = "Playwright + Pytest automation framework"
authors = ["Automation Team <qa@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
playwright = ">=1.40,<2.0"
pytest = ">=7.0,<9.0"
allure-pytest = ">=2.13,<3.0"

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"